inventory_control/requirements.txt
aiodns==2.0.0
aiohttp==3.7.3
aioredis==1.3.1
aniso8601==8.1.0
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.3.1
astroid==2.5.3
asttokens==2.0.5
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
autopep8==1.5.6
Babel==2.9.0
backcall==0.2.0
base32-lib==1.0.2
bcrypt==3.2.0
beautifulsoup4==4.9.3
bidict==0.21.2
bitstring==3.1.7
blinker==1.4
cachetools==4.2.1
certifi==2019.11.28
cffi==1.14.4
channels==2.4.0
channels-redis==3.2.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
constantly==15.1.0
cryptography==3.3.1
cupshelpers==1.0
Cython==0.29.22
daphne==2.5.0
dateparser==1.0.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.1.3
django-better-admin-arrayfield==1.4.2
django-bootstrap-form==3.4
django-bootstrap4==2.3.1
django-countries==7.1
django-crispy-forms==1.10.0
django-debug-toolbar==3.2
django-formtools==2.2
django-otp==1.0.2
django-otp-yubikey==1.0.0
django-phonenumber-field==3.0.1
django-redis==4.12.1
django-tawkto==0.3
django-two-factor-auth==1.13
django-user-sessions==1.7.1
docutils==0.17.1
ecdsa==0.16.1
entrypoints==0.3
esptool==3.0
et-xmlfile==1.1.0
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-RESTful==0.3.8
Flask-SocketIO==5.0.1
gevent==21.1.1
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.27.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.53.0
greenlet==1.0.0
gunicorn==20.0.4
h11==0.12.0
hiredis==1.1.0
httplib2==0.19.0
hyperlink==21.0.0
idna==2.8
incremental==17.5.0
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
isort==5.8.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.2
jupyter-client==6.1.11
jupyter-core==4.7.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.6.0
louis==3.12.0
macaroonbakery==1.3.1
mailjet-rest==1.3.4
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.2
multidict==5.1.0
mypy==0.812
mypy-extensions==0.4.3
mysql-connector-python==8.0.23
netifaces==0.10.4
notify2==0.3
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.1.48
openpyxl==3.0.7
packaging==20.9
pandas==1.2.0
paramiko==2.7.2
parso==0.8.1
pbkdf2==1.3
pexpect==4.6.0
phonenumbers==8.12.16
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==3.0.17
protobuf==3.15.3
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycares==3.1.1
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.9.9
pycups==1.9.73
Pygments==2.8.1
PyGObject==3.36.0
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.7.4
pymacaroons==0.13.0
pymongo==3.11.2
PyNaCl==1.3.0
pyOpenSSL==20.0.1
pyotp==2.4.1
pyparsing==2.4.7
pyRFC3339==1.1
pyserial==3.5
pytesseract==0.3.7
python-apt==2.0.0+ubuntu0.20.4.5
python-binance==1.0.7
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-engineio==4.0.0
python-socketio==5.0.4
python-telegram-bot==13.1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.0.3
qrcode==6.1
redis==3.5.3
redis-server==6.0.9
reedsolo==1.5.4
regex==2020.11.13
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7.2
Rx==3.1.1
SecretStorage==2.3.1
Send2Trash==1.5.0
service-identity==18.1.0
simplejson==3.16.0
singleton-decorator==1.0.0
six==1.14.0
soupsieve==2.1
sqlparse==0.4.1
systemd-python==234
ta==0.7.0
tesseract==0.1.3
tesseract-ocr==0.0.1
thonny==3.3.10
toml==0.10.2
tornado==6.1
traitlets==5.0.5
twilio==6.51.0
Twisted==20.3.0
txaio==20.12.1
typed-ast==1.4.3
typing-extensions==3.7.4.3
tzlocal==2.1
ubuntu-advantage-tools==27.0
ubuntu-drivers-common==0.0.0
ufw==0.36
ujson==4.0.1
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.25.8
uvicorn==0.13.3
virtualenv==20.4.2
wadllib==1.3.3
wcwidth==0.2.5
websockets==8.1
Werkzeug==1.0.1
wifi==0.3.8
wrapt==1.12.1
xkit==0.0.0
xlrd==1.2.0
yarl==1.1.0
YubiOTP==1.0.0
zope.event==4.5.0
zope.interface==5.2.0
